+++ update-server/tests/common/conftest.py
@@ -1,12 +1,19 @@
+import lzma
 import os
 import subprocess
 import re
+import zipfile
+from unittest import mock
+from unittest.mock import MagicMock
 
 import pytest
 
-from otupdate import buildroot
+from otupdate import buildroot, common
 
 from otupdate import openembedded
+from otupdate.common.update_actions import Partition
+from otupdate.openembedded import PartitionManager
+from tests.common.config import FakeRootPartElem
 
 HERE = os.path.abspath(os.path.dirname(__file__))
 one_up = os.path.abspath(os.path.join(__file__, "../../"))
@@ -26,68 +33,189 @@ async def test_cli(aiohttp_client, loop, otupdate_config, request):
         loop=loop,
     )
     client = await loop.create_task(aiohttp_client(app))
-    return client
+    return client, cli_client_pkg.__name__
 
 
-br_files = dict(
-    [
-        ("ROOTFS", "rootfs.ext4"),
-        ("ROOTFS_HASH", "rootfs.ext4.hash"),
-        ("ROOTFS_SIG", "rootfs.ext4.hash.sig"),
+@pytest.fixture
+def downloaded_update_file_consolidated(request, extracted_update_file_consolidated):
+    """
+    Return the path to a zipped update file
+
+    To exclude files, mark with ``exclude_rootfs_ext4``,
+    ``exclude_rootfs_ext4_hash``, ``exclude_rootfs_ext4_hash_sig``.
+
+    This uses :py:meth:`extracted_update_file` to generate the contents, so
+    marks that fixture understands can be used when requesting this fixture
+
+    Can also be used by tests that will upload it to a test server, since
+    when the test server boots its download path will be somewhere else
+    """
+    zip_path_arr = []
+    list_of_update_files = [
+        (
+            "rootfs.xz",
+            "rootfs.xz.sha256",
+            "rootfs.xz.hash.sig",
+            "tmp_uncomp_xz_hash_path",
+            "ot3-system.zip",
+        ),
+        (
+            "rootfs.ext4",
+            "rootfs.ext4.hash",
+            "rootfs.ext4.hash.sig",
+            "tmp_uncomp_xz_hash_path",
+            "ot2-system.zip",
+        ),
     ]
-)
-oe_files = dict(
-    [
-        ("ROOTFS", "rootfs.ext4"),
-        ("ROOTFS_HASH", "rootfs.ext4.hash"),
-        ("ROOTFS_SIG", "rootfs.ext4.hash.sig"),
-    ]
-)
+    for index, (rootfs, sha256, sig, xz_hash, pkg) in enumerate(list_of_update_files):
+        rootfs_path = os.path.join(extracted_update_file_consolidated[index], rootfs)
+        hash_path = os.path.join(extracted_update_file_consolidated[index], sha256)
+        sig_path = os.path.join(extracted_update_file_consolidated[index], sig)
+        xz_hash_path = os.path.join(extracted_update_file_consolidated[index], xz_hash)
+        zip_path = os.path.join(extracted_update_file_consolidated[index], pkg)
+        with zipfile.ZipFile(zip_path, "w") as zf:
+            if not request.node.get_closest_marker("exclude_rootfs_ext4"):
+                zf.write(rootfs_path, rootfs)
+            if not request.node.get_closest_marker("exclude_rootfs_ext4_hash"):
+                zf.write(hash_path, sha256)
+            if not request.node.get_closest_marker("exclude_rootfs_ext4_hash_sig"):
+                zf.write(sig_path, sig)
+            zf.write(xz_hash_path, xz_hash)
+            zip_path_arr.append(zip_path)
+        os.unlink(rootfs_path)
+        os.unlink(hash_path)
+        os.unlink(sig_path)
+
+    return zip_path_arr
 
 
-@pytest.fixture(params=[br_files, oe_files])
-def extracted_update_file(request, tmpdir):
+def write_fake_rootfs(
+    rootfs_name: str, rootfs_path: str, rootfs_contents: bytes, uncomp_xz_path: str
+) -> str:
+    if rootfs_name == "rootfs.xz":
+        with lzma.open(rootfs_path, "w") as f:
+            f.write(rootfs_contents)
+        with lzma.open(rootfs_path, "rb") as fsrc, open(uncomp_xz_path, "wb") as fdst:
+            while True:
+                chunk = fsrc.read(1024)
+                fdst.write(chunk)
+                if len(chunk) != 1024:
+                    break
+        return uncomp_xz_path
+    else:
+        with open(rootfs_path, "wb") as rfs:
+            rfs.write(rootfs_contents)
+            return rootfs_path
+
+
+def gen_hash_val_direct(rfs_path: str) -> str:
+    try:
+        shasum_out = subprocess.check_output(
+            [
+                "shasum",
+                "-a",
+                "256",
+                rfs_path,
+            ]
+        )
+        return shasum_out
+    except (subprocess.CalledProcessError, FileNotFoundError):
+        pytest.skip("no shasum invokeable on command line")
+
+
+@pytest.fixture
+def extracted_update_file_consolidated(request, tmpdir):
     """
     Return the path to a dir containing an unzipped update file.
 
     To make a bad hash, mark with ``bad_hash``. To make a bad
     signature, mark with ``bad_sig``.
     """
-    files = request.param
-    rootfs_path = os.path.join(tmpdir, files["ROOTFS"])
-    hash_path = os.path.join(tmpdir, files["ROOTFS_HASH"])
-    sig_path = os.path.join(tmpdir, files["ROOTFS_SIG"])
-    rootfs_contents = os.urandom(100000)
-    with open(rootfs_path, "wb") as rfs:
-        rfs.write(rootfs_contents)
-    if request.node.get_closest_marker("bad_hash"):
-        hashval = b"0oas0ajcs0asd0asjc0ans0d9ajsd0ian0s9djas"
-    else:
-        try:
-            shasum_out = subprocess.check_output(["shasum", "-a", "256", rootfs_path])
-        except (subprocess.CalledProcessError, FileNotFoundError):
-            pytest.skip("no shasum invokeable on command line")
-        hashval = re.match(b"^([a-z0-9]+) ", shasum_out).group(1)
-    with open(hash_path, "wb") as rfsh:
-        rfsh.write(hashval)
-    if not request.node.get_closest_marker("bad_sig"):
-        try:
-            subprocess.check_output(["openssl", "version"])
-        except (subprocess.CalledProcessError, FileNotFoundError):
-            pytest.skip("requires openssl binary to be installed")
-        subprocess.check_call(
-            [
-                "openssl",
-                "dgst",
-                "-sha256",
-                "-sign",
-                os.path.join(one_up, "ot-update-server-unit-tests.key"),
-                "-out",
-                sig_path,
-                hash_path,
-            ]
-        )
-    else:
-        with open(sig_path, "wb") as sigfile:
-            sigfile.write(os.urandom(256))
-    return tmpdir
+    extracted_files_dir_path_arr = []
+    list_of_extracted_files = [
+        (
+            "rootfs.xz",
+            "rootfs.xz.sha256",
+            "rootfs.xz.hash.sig",
+        ),
+        (
+            "rootfs.ext4",
+            "rootfs.ext4.hash",
+            "rootfs.ext4.hash.sig",
+        ),
+    ]
+
+    for (rootfs, sha256, sig) in list_of_extracted_files:
+        rootfs_path = os.path.join(tmpdir, rootfs)
+        hash_path = os.path.join(tmpdir, sha256)
+        uncomp_xz_hash_path = os.path.join(tmpdir, "tmp_uncomp_xz_hash_path")
+        sig_path = os.path.join(tmpdir, sig)
+        uncomp_xz_path = os.path.join(tmpdir, "tmp_uncomp")
+        rootfs_contents = os.urandom(100000)
+        write_fake_rootfs(rootfs, rootfs_path, rootfs_contents, uncomp_xz_path)
+        if request.node.get_closest_marker("bad_hash"):
+            hashval = b"0oas0ajcs0asd0asjc0ans0d9ajsd0ian0s9djas"
+        else:
+
+            hashval = re.match(
+                b"^([a-z0-9]+) ",
+                gen_hash_val_direct(rootfs_path),
+            ).group(1)
+            hashval2 = re.match(
+                b"^([a-z0-9]+) ",
+                gen_hash_val_direct(uncomp_xz_path),
+            ).group(1)
+        with open(hash_path, "wb") as rfsh:
+            rfsh.write(hashval)
+        with open(uncomp_xz_hash_path, "wb") as rfsh:
+            rfsh.write(hashval2)
+        if not request.node.get_closest_marker("bad_sig"):
+            try:
+                subprocess.check_output(["openssl", "version"])
+            except (subprocess.CalledProcessError, FileNotFoundError):
+                pytest.skip("requires openssl binary to be installed")
+            subprocess.check_call(
+                [
+                    "openssl",
+                    "dgst",
+                    "-sha256",
+                    "-sign",
+                    os.path.join(one_up, "ot-update-server-unit-tests.key"),
+                    "-out",
+                    sig_path,
+                    hash_path,
+                ]
+            )
+        else:
+            with open(sig_path, "wb") as sigfile:
+                sigfile.write(os.urandom(256))
+        extracted_files_dir_path_arr.append(tmpdir)
+    return extracted_files_dir_path_arr
+
+
+@pytest.fixture
+def testing_partition(monkeypatch, tmpdir):
+    partfile = os.path.join(tmpdir, "fake-partition")
+    find_unused = mock.Mock()
+    monkeypatch.setattr(buildroot.update_actions, "_find_unused_partition", find_unused)
+
+    find_unused.return_value = FakeRootPartElem(
+        "TWO", common.update_actions.Partition(2, partfile)
+    )
+    return partfile
+
+
+@pytest.fixture
+def mock_partition_manager_valid_switch(tmpdir) -> MagicMock:
+    """Mock Partition Manager."""
+    partfile = os.path.join(tmpdir, "fake-partition")
+    mock_part = MagicMock(spec=PartitionManager)
+    mock_part.find_unused_partition.return_value = Partition(2, partfile)
+    mock_part.switch_partition.return_value = Partition(2, partfile)
+    mock_part.resize_partition.return_value = True
+    mock_part.mount_fs.return_value = True
+    mock_part.umount_fs.return_value = True
+
+    mock_part.mountpoint_root.return_value = "/mnt"
+
+    return mock_part
